# openapi.yaml
openapi: 3.1.0
info:
  title: RFO Protocol API
  version: 1.0.0
  description: |
    ## RFO (Research For Optimization) Protocol
    
    A next-generation application-layer network protocol designed for AI web-scale agents.
    RFO enables structured, verifiable, and efficient content delivery between origin servers
    and AI agents through cryptographic site identification, quality scoring, and
    token-optimized payloads.
    
    ### Key Concepts
    - **Handshake**: Bidirectional protocol where the agent provides a domain URL and the engine
      compiles, verifies, and returns structured `.doc` or `.mdoc` payloads.
    - **Site ID**: Cryptographically derived, deterministic identity tied to domain + secret + time window.
    - **Quality Score**: Automated 0–100 scoring of content structure, verifiability, and AEO readiness.
    - **Payload Types**:
      - `.doc` — Full deep-knowledge layout (markdown, data tables, verification signature).
      - `.mdoc` — Token-optimized mini-document (< 1,500 tokens) for LLM context windows.
    - **Streaming**: Binary chunked transfer for large payloads with checksums.
    - **WebSocket**: Real-time pub/sub for domain update notifications.
    - **Capability Negotiation**: Protocol version and encoding negotiation (JSON / MessagePack).
    
    ### Authentication
    All protected endpoints require an `X-API-Key` header. Public endpoints (`/rfo/health`,
    `/rfo/capabilities`, `/rfo/negotiate`) do not require authentication.
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for client authentication. Obtain keys from your RFO deployment administrator.
        Format: `X-API-Key: <your-api-key>`
  schemas:
...
    CapabilityRequest:
      type: object
      required: [supported_encodings, supported_features, protocol_version]
      properties:
        supported_encodings:
          type: array
          items:
            type: string
            enum: [application/json, application/msgpack]
        supported_features:
          type: array
          items:
            type: string
        protocol_version:
          type: string
          example: "1.0.0"
    CapabilityResponse:
      type: object
      required: [negotiated_encoding, supported_features, protocol_version, server_capabilities]
      properties:
        negotiated_encoding:
          type: string
          enum: [application/json, application/msgpack]
        supported_features:
          type: array
          items:
            type: string
        protocol_version:
          type: string
        server_capabilities:
          type: array
          items:
            type: string

    # ── WebSocket ─────────────────────────────────────────────────────────────
...
paths:
...
  /rfo/negotiate:
    post:
      tags: [Protocol]
      summary: Negotiate capabilities
      description: |
        Public endpoint. Client sends its supported encodings and features;
        server responds with the negotiated encoding and common feature set.
      operationId: negotiate
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CapabilityRequest'
      responses:
        '200':
          description: Negotiated capabilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'

  # ── Handshake ───────────────────────────────────────────────────────────────
...
security:
  - ApiKeyAuth: []